Added Passport.js and SendGrid Functionality.
Passport authenticates the user with google oauth2 and redirects to protected page called success. Where The user can subscribe to news letter, Which adds the user to a database of users who receive news letter. When a news letter is sent a record is created as date in its name holding the users who received the email.